requirements.txt
langchain==0.3.10
langchain-anthropic==0.3.10
langchain-openai==0.3.10
langchain-core==0.3.51
PyPDF2==3.0.1
python-dotenv==1.1.0
anthropic==0.49.0
openai==1.72.0